Allow building from git without gnulib-tool
Since we don't want to keep Gnulib translations in revision control,
we need gnulib-tool on $PATH when generating the build system in
order to fetch them, but it's unnecessarily awkward to require
everyone to have this just in order to build man-db from git.  Allow
building without this, although it is still required for "make
dist".

* configure.ac: No longer issue an error if gnulib/po/Makefile.in.in
is missing.  Instead, set the Automake conditional HAVE_GNULIB_PO to
true if and only if gnulib/po/POTFILES.in exists, and only create
gnulib/po/Makefile.in if gnulib/po/Makefile.in.in exists.
* Makefile.am (SUBDIRS): Only include gnulib/po if HAVE_GNULIB_PO is
true.