Separate cache owner from --enable-setuid option
It's useful to have a notion of the cache owner even when man is not
installed setuid.  --enable-setuid no longer takes an argument, and the
owner is now set by the --enable-cache-owner option instead.

* m4/man-arg-cache-owner.m4: New file.
* m4/man-arg-setuid.m4: Stop accepting an argument.  Only set man_mode,
not man_owner.
* configure.ac: Call MAN_ARG_CACHE_OWNER.
* src/Makefile.am (install-exec-hook): Only chown man and mandb if
man_mode is 4755 (as well as the existing test for man_owner being
non-empty).