Add fallback pager if the compile-time default is not executable
A problem with man-db's man is that in the case of the user giving no
configuration via conf files, argv, or environment variables; man
defaults to less as pager (PAGER); but less may not be present on the
system. Sure, other pagers may be selected in aforementioned ways, but
then the defaults are overridden, making that unsuitable for
install-time configuration.

This patch makes man check (if that becomes relevant) if PAGER is
executable, further defaulting to cat (which is basically ubiquitous,
being in original Unix, POSIX, and GNU Coreutils) if it is not. Thus
the poor beginner Unix users without less installed will be able to
get man pages.

* src/man.c (sh_lang_first_word): New function.
(main): Skip configured pager if it is not executable.
* man/man1/man.man1 (Controlling formatted output, ENVIRONMENT):
Document fallback to cat.
* man/replace.sin.in: Substitute %cat%.