sandbox: Work around ESET File Security
This is a proprietary antivirus program, so this is only a best guess
from strace output.  The choices are to disable the sandbox entirely or
to allow a few socket-related system calls if this antivirus program is
detected, and the latter is probably slightly better.

Reported by John Sivak.

* lib/sandbox.c (make_seccomp_filter): If LD_PRELOAD contains the
substring "/libesets_pac.so", then allow some socket-related system
calls so that the preload wrapper can talk to its daemon.
* NEWS: Document this.