Define an access(2) wrapper with clearer semantics
As usual for system calls, access(2) returns zero on success.  However,
I generally think of it as "can we access this file in this way", where
boolean semantics would be more convenient, and find it too easy to
invert logic by accident when using the system call directly.  Define a
CAN_ACCESS wrapper with boolean semantics.

* include/manconfig.h.in (CAN_ACCESS): New macro.
* lib/tempfile.c (path_search): Use CAN_ACCESS.
* src/catman.c (check_access): Likewise.
* src/filenames.c (make_filename): Likewise.
* src/man.c (make_roff_command, display): Likewise.
* src/ult_src.c (find_include): Likewise.
* src/whatis.c (use_grep): Likewise.