sandbox: Work around Microsoft SCEP
This is a proprietary antivirus program and I've only been able to guess
at how to handle it.  Note that it is no longer supported by Microsoft
and so users should probably replace it with something else, but I still
want to minimise the number of support requests I get related to it.

* lib/sandbox.c (make_seccomp_filter): If libscep_pac.so is preloaded,
then allow some system calls related to sockets and System V message
queues.
* NEWS: Document this.