man(1): Minor typographic changes
Fixes Debian bug #892230.

* man/man1/man.man1 (DESCRIPTION, HISTORY): Protect "." with "\&" if it
is followed by a space and does not mean the end of a sentence.
(EXAMPLES): Use a macro to change to the italic font instead of \fI if
possible.  The macros have the italic corrections.
(EXAMPLES, Controlling formatted output, Getting help, SEE ALSO): Change
.BR/.IR to .B/.I if there is only one argument, or else separate the
second argument from the first if needed.
(DEFAULTS): Begin each sentence on a new line.
(FILES): Add a space around "|" to increase readability.