apropos(1): Minor typographic changes
Fixes Debian bug #892421.

* man/man1/apropos.man1 (DESCRIPTION): Use "\e" to print the escape
character instead of "\\" (which is interpreted in copy mode).
(OPTIONS): Change .BR to .B if there is only one argument.
(FILES): Add a space around "|" to increase readability.
(AUTHOR): Protect "." with "\&" if it is followed by a space and does
not mean the end of a sentence.