Make gxditview work with seccomp sandbox
"man -X" and friends used to work by running groff with -X, causing it
to run gxditview itself.  This didn't work with the seccomp sandbox: the
sorts of system calls that a text processor needs to make are not very
much related to the sorts of system calls that an X client needs to
make.

Instead, run groff with -X -Z to pass the appropriate options to troff
but stop it from running gxditview, and then run gxditview ourselves
without sandboxing.  This also lets us set a better window title.

* src/man.c (make_roff_command): In gxditview mode, run groff with -Z as
well as -X.
(make_display_command): In gxditview mode, add an appropriate gxditview
process.
(display): In gxditview mode, call the ordinary "display formatted page"
logic, without saving a cat page.