Remove "Type" column from DB table in mandb(8)
The database types table in mandb(8) has always been a bit
dubiously-placed; it's really more about how man-db is configured, and
as such makes more sense in manual/db.me (which indeed has an extended
version of the same table).  It causes particular problems in translated
versions of this page, where it's very easy for it to become too wide
and cause "make check" failures.

A proper fix for this would probably require updating existing
translations, but an easy improvement is to remove the "Type" column: it
has negligible relevance here, and removing it makes it much easier for
translated versions to stay within page width limits.

* man/man8/mandb.man8 (DATABASE CACHES): Remove "Type" column from
table.