mandb: Fix some details of error handling
* src/check_mandirs.c (testmandirs): Return -1 rather than -errno on
failure to create the database.
(create_db, update_db): Only call update_db_time if amount > 0, rather
than if it is non-zero (since it may be negative to indicate an error,
and updating the database time doesn't make sense in that case).
(update_db): Return -1 rather than EOF.
* src/mandb.c (update_db_wrapper): Check whether amount is non-zero
rather than whether it is unequal to EOF.