Stop GitLab from getting confused about licences
GitLab uses https://github.com/licensee/licensee to detect the project's
licence, and if that finds multiple matching files then it throws up its
hands and refuses to make a licence assertion, with no apparent way to
override it.

To avoid this problem, move the secondary licence files (those that
apply only to some files, rather than to the project as a whole) back
into a subdirectory, so that `licensee` doesn't notice them.

* COPYING.GPLv2: Move to ...
* docs/COPYING.GPLv2: ... here.  Update all references.
* COPYING.LIB: Move to ...
* docs/COPYING.LIB: ... here.  Update all references.
* Makefile.am (dist_noinst_DATA): Remove COPYING.GPLv2.
* docs/Makefile.am (dist_noinst_DATA): Add COPYING.GPLv2 and
COPYING.LIB.