Stop using the addr field of struct mandata
The `addr` field of `struct mandata` is just intended as an internal
buffer, but some functions were extracting information from it in
complicated ways, which made it difficult to refactor the memory
allocation here.  Use different approaches.

* src/check_mandirs (test_manfile): Calculate `len` using `comp_info`.
This duplicates some of what `filename_info` does, but it's just some
cheap string manipulation, and it expresses what we want in a less
confusing way (the length of the file name with any compression
extension removed).
* src/descriptions_store.c (store_descriptions): Check whether `path` is
a prefix of `trace_name` rather than of `trace_info->addr`.
`trace_info->addr` is always itself a prefix of `trace_name`, so if
`path` is not a prefix of `trace_name` then it can't be a prefix of
`trace_info->addr`.