Fix stored ID for links
Database entries for links were often incorrectly stored as `ULT_MAN`,
depending on the exact order in which pages were scanned.  They are now
stored more consistently as `SO_MAN`.

* src/descriptions_store.c (store_descriptions): Override the ID from
`ULT_MAN` to `SO_MAN` if the name is not the last entry in the trace.
(Previously we overrode from `SO_MAN` to `ULT_MAN` if the name was the
last entry in the trace, but not the converse.)
* NEWS.md: Document this.