Pick a more consistent name for whatis targets
We don't want the targets of whatis entries in the database to depend on
the order in which files were scanned.  Prefer the first name in the
trace out of those that have the best ID.

* src/descriptions_store.c (store_descriptions): Defer adding names that
don't appear in the ultimate source trace (i.e. whatis entries) until we
know the best name to use as the target, and use that best name as the
pointer for each of them.