Fix section/extension for links in database
In some cases this resulted in spurious database entries.  For example,
if `man5/inetd.conf.5` was a symlink to `man8/inetd.8` and that page had
a `NAME` entry for `inetd`, then as well as entries for `inetd` in
section 8 and `inetd.conf` in section 5, we would also store a spurious
entry for `inetd` in section 5.  We no longer do that.

* src/descriptions_store.c (store_descriptions): For names corresponding
to entries in the ultimate source trace and not to the base name of the
file, use the section and extension of the trace entry rather than of
the ultimate source.