Extend database entry replacement rules
These rules were previously ambiguous such that the targets of whatis
entries depended on the order in which files were scanned.

* libdb/db_store.c (enum replace_action): Add.
(replace_if_necessary): Extend rules for greater stability.  We now also
take mtimes and pointers into account, as well as including a case for
when the new ID is weaker than the old one (there was already a case for
when it was stronger).
(dbstore): Use `replace_if_necessary` when considering replacing multi
keys.