ult_src: Remove recursion
The use of recursion for `.so` handling in `ult_src` is somewhat
inconvenient, and it's relatively easy to convert it to an iterative
implementation.

There should be no functional change from this commit, except to some
details of debug messages.

* src/ult_src.c (find_include_directive): New function, extracted from
`ult_src`.
(ult_src): Handle `.so` links using a loop rather than recursion.  Check
for compression extensions towards the end of the loop rather than at
the beginning, to avoid nonexistent files showing up in the trace.