Further improve name selection for whatis targets
In commit 9aff3c30b0, I used the rule "first name in the trace out of
those that have the best ID".  However, a better rule turns out to be to
pick the last entry in the trace that has the same section and extension
as the starting page.  This is already guaranteed to have at least as
good an ID as the starting page, and since whatis pointers don't include
the target section and extension we need to ensure that they don't cross
sections/extensions; otherwise we just end up with database entries that
man(1) refuses to use, which isn't particularly useful.

* src/descriptions_store.c (store_descriptions): Parse and cache all the
filenames in the trace at the start.  Use the last entry in the trace
with the same section/extension as the starting page as the pointer for
`WHATIS_*` entries.