test_manfile: Remove "already exists" check
This produced inconsistent results in some unusual cases.  For example,
if `/usr/share/man/man5/inetd.conf.5.gz` was a symlink to
`/usr/share/man/man8/inetd.8.gz` and had `NAME` entries for both `inetd`
and `inetd.conf`, then if `inetd.8.gz` was scanned first we would
include a `WHATIS_MAN` entry in section 8 for `inetd.conf`, while if
`inetd.conf.5.gz` was scanned first we'd skip that due to hitting the
already-exists check when scanning `inetd.8.gz`.

This was originally a performance optimization.  Now that we cache the
results of `ult_src` and `find_name`, this optimization no longer helps
us much, so it can go.

* src/check_mandirs.c (test_manfile): Don't return early if the page
already exists with matching details.
* NEWS.md: Document this.