Include .lf corrections in debugging output
The pipeline command names in `PIPELINE_DEBUG` output are deliberately
abbreviated in the cases where the command is inserting data into the
`*roff` input stream, in an attempt to convey the essential parts
without overwhelming the reader.  However, it probably does make sense
to include `.lf` corrections there, since if you're relying on the
debugging output to construct a manual shell pipeline when debugging an
issue with a page, you'll want the `.lf` corrections so that line
numbers still come out correctly.

* src/man.c (add_roff_line_length, display): Include `echo .lf 1` in
command names where appropriate.