lexgrog.1: Fix thinko in my recent contribution
In commit bbf7701c4f, 19 July, I got the sense of the tests that
governed the macro definitions backwards.  I fooled myself because I use
a more complex test in groff's an-ext.tmac file (written in a
non-intuitive way due to AT&T troff limitations) so that it's easy to
run groff's automated tests on them.