man: simplify the right-margin condition
As it stands, this is
  ⌊line_length*(39/40)⌋ > line_length - 2
which works out to
  line_length ≤ 40
Is that obvious? No (I analysed it to line_length < 80),
but the way the result is used makes it equivalent to
  line_length ≤ 80
anyway.

This currently blames to commit cac2ef84fe626f112de7537c20dfe29c35a71b96
("Update default line length in line with groff 1.18.") dated
2002-07-06 20:30:34 +0000(!) with a diff of
  -       if (!troff && (line_length < 66 || line_length > 80))
  -               return line_length * 9 / 10;
  +       /* groff >= 1.18 defaults to 78. */
  +       if (!troff && line_length != 80) {
  +               int length = line_length * 39 / 40;
  +               if (length > line_length - 2)
  +                       return line_length - 2;
  +               else
  +                       return length;

NFC