man: match the display width to the configured width
Currently, the line length is:
  if w ≤ 40: w - 2
  else     : ⌊w*(39/40)⌋
i.e.
  if w ≤ 40: w - 2
  else     : w - ⌈w/40⌉
(this is equivalent to a ≤80 condition).

This was originally originally introduced in commit
f139e8273d5f3754f0d0dacb890e515b30c2d28b ("Add terminal line length
 detection to man.") from 2001-06-28, which did
  if w ∈ [66, 80]: 80    (via default)
  else           : ⌊.9w⌋
then tracked the default in
commit cac2ef84fe626f112de7537c20dfe29c35a71b96 ("Update default
 line length in line with groff 1.18.") from 2002-07-06,
which updates it to the current formula,
then commit 39cdee4ada59d711d066628ec827508070d90a70 ("Always pass a
 line length to nroff") from 2023-08-06 removes any defaulting.

This is ideologically quite weird, because this behaves like a
stochastic best-effort fit so there is always a blank margin, growing
with size, derived from a 21-year-old default which was only relevant
because catpages were vaguely relevant, even though we perfectly
know how large we need to display the manual by virtue of perfectly
knowing the size of the teletype! There is no such margin on the
left, and why would there be?

Fixes Debian bug #1059537.