Further improve table formatting in description of --ascii
Fixes https://bugs.debian.org/1073530, again.

The use of text blocks here was an attempt (in
6000c56fb621b454185b04d0d718ede9a87eadf4) to make this page more robust
against the possibility of translations that were long enough to cause
the table to be overly wide otherwise.  However, in the 13 years since I
added them, no language has produced long enough translations for this
to be a problem, and this defence against a hypothetical problem causes
real formatting issues because `tbl` formats text blocks in a diversion
whose width is calculated in advance and is difficult to set
appropriately.

If a translation comes along in future that pushes the column limits
harder, then one option would be to drop the "Octal" column, which
arguably provides an excessive level of detail.

* man/man1/man.man1 (Controlling formatted output): In the description
of `--ascii`, abandon text blocks and drop workarounds that attempted to
improve their formatting.